libc: Fix prototype so libc builds again...

+++ lib/libc/posix/ftruncate.c
@@ -5,7 +5,7 @@
 #include <syscalls.h>
 #include <_posix/fdmap.h>
 
-off_t ftruncate( int fd, off_t length )
+int ftruncate( int fd, off_t length )
 {
 	void* handle = fdmap_deref(fd);
 	if (handle == NULL)
